requirements.txt
opencv-python
albumentations
tqdm
requests
matplotlib
seaborn
wandb
segmentation_models_pytorch